Refactored campaign; random picker; some CampaignList views
